requirements.txt
asgiref==3.8.1
Django==5.0.6
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
djangorestframework-simplejwt
dj-rest-auth[with_social]
Pillow